pyproject.toml
[project]
name = "azure-ai-agent-service-quickstart"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-ai-projects>=1.0.0b12",
    "azure-ai-agents>=1.1.0b2",
    "azure-identity>=1.23.0",
    "python-dotenv>=1.1.0",
]
